Assembler: move reqvm::macro to its own file
